// frontend/src/features/actividades/asignar_estudio_titulos/models/asignar_estudio_titulos.ts
export interface AsignarEstudioTitulos {
  id_asignar_estudio_titulos: number;
  id_expediente: number;
  id_actividad: string;
  abogado: string;
  observaciones: string;
  is_active: boolean;
  row_status: boolean;
  created_by: number;
  created_date: string;
  modified_by?: number | null;
  modified_date?: string | null;
}